error in navbar responsive

+++ server/src/services/data/data.service.ts
@@ -12,7 +12,7 @@ import {
   Notification,
   NotificationType,
 } from 'src/core/entities/notification.entity';
-import { fullImagePath } from '../helpers/image-storage';
+import { fullImagePath, isFileExtensionSafe } from '../helpers/image-storage';
 import {
   Conversation,
   ConversationType,
@@ -250,11 +250,15 @@ export class DataService {
     myId: number,
   ) {
     if (file || body.username) {
-      if (file)
-        await this.usersService.updateAvatar(
-          myId,
-          fullImagePath(file.filename),
-        );
+      if (file) {
+        const check = await isFileExtensionSafe(file.path);
+        if (check == true) {
+          await this.usersService.updateAvatar(
+            myId,
+            fullImagePath(file.filename),
+          );
+        } else console.log('file is not safe');
+      }
       if (body.username)
         await this.usersService.updateUsername(myId, body.username);
       return { message: 'Profile is updated' };
@@ -263,6 +267,8 @@ export class DataService {
 
   async updateAvatar(file: Express.Multer.File, req) {
     if (!file) return { message: 'avatar not updated' };
+    const check = await isFileExtensionSafe(file.path);
+    if (check == false) return { message: 'file is not safe' };
     await this.usersService.updateAvatar(
       req.user.id,
       fullImagePath(file.filename),
@@ -404,12 +410,8 @@ export class DataService {
     let arr = [];
     await Promise.all(
       allChannels.map(async (newChannel) => {
-        let status: string;
         const channelUser = await this.findChannelUser(newChannel.id, id);
         if (!channelUser) {
-          console.log('amine sba3');
-          if (!channelUser) status = 'join';
-          else if (channelUser.ban === true) status = 'blocked';
           const object = {
             id: newChannel.id,
             name: newChannel.name,
@@ -417,7 +419,7 @@ export class DataService {
             createdAt: format(newChannel.createdAt, 'MMMM dd,yyyy'),
             members: newChannel.members,
             owner: newChannel.owner,
-            status: status,
+            status: 'join',
           };
           arr.push(object);
         }
@@ -436,9 +438,11 @@ export class DataService {
           (element.ban == true || element.mute == true) &&
           element.period > 0
         ) {
+          console.log(element.ban);
           const firstDate = moment(element.timeOfOperation);
           const nowDate = moment(new Date());
           const interval = Math.abs(firstDate.diff(nowDate, 'minutes'));
+          console.log(interval);
           if (interval >= element.period) {
             if (element.ban == true)
               await this.removeBanUserInChannel(element.channel.id, id);
@@ -617,10 +621,16 @@ export class DataService {
     const user: User = await this.usersService.findOneById(me);
     const channelUser: ChannelUser =
       await this.channelUserService.findbyChannelAndUser(channel, user);
-    if (!channelUser)
-      if (!channel)
-        return { status: 500, message: 'U dont have acces to this channel' };
-    const object = { ...channel, myRole: channelUser.userType, mute: channelUser.mute };
+    if (!channelUser) {
+      console.log('findMyChannel');
+      throw UnauthorizedException;
+      // return { status: 500, message: 'U dont have acces to this channel' };
+    }
+    const object = {
+      ...channel,
+      myRole: channelUser.userType,
+      mute: channelUser.mute,
+    };
     return object;
   }
 
@@ -650,6 +660,7 @@ export class DataService {
   }
 
   async listUsersOfChannelWitouhtMe(channelId: number) {
+    console.log('-------------------');
     const newChannel: Channel = await this.channelService.findChannelById(
       channelId,
     );
@@ -684,7 +695,6 @@ export class DataService {
     }
     all.adminsNumber = all.admins.length;
     all.usersNumber = all.users.length;
-    // console.log(all);
     return all;
   }
 
@@ -725,31 +735,43 @@ export class DataService {
     const newChannel: Channel = await this.channelService.findChannelById(
       channelId,
     );
-    // console.log(newChannel);
     const newUser: User = await this.usersService.findOneById(myId);
-    // console.log(newUser);
     const chanelUser = await this.channelUserService.findbyChannelAndUser(
       newChannel,
       newUser,
     );
-    // console.log(chanelUser);
     return chanelUser;
   }
 
   async changeUserToBeAdminInChannel(channelId: number, myId: number) {
+    const channel: Channel = await this.channelService.findChannelById(
+      channelId,
+    );
+    if (!channel) return { status: 500, message: 'Channel not found' };
     const chanelUser = await this.findChannelUser(channelId, myId);
+    if (!chanelUser)
+      return { status: 500, message: 'This user is not found in this channel' };
     return await this.channelUserService.updateToBeAdmin(chanelUser.id);
   }
 
   async changeAdminToBeUserInChannel(channelId: number, myId: number) {
+    const channel: Channel = await this.channelService.findChannelById(
+      channelId,
+    );
+    if (!channel) return { status: 500, message: 'Channel not found' };
     const chanelUser = await this.findChannelUser(channelId, myId);
+    if (!chanelUser)
+      return { status: 500, message: 'This user is not found in this channel' };
     return await this.channelUserService.updateToBeUser(chanelUser.id);
   }
 
   async kickUserFormChannelConversation(channelId: number, myId: number) {
+    const channel: Channel = await this.channelService.findChannelById(
+      channelId,
+    );
+    if (!channel) return { status: 500, message: 'Channel not found' };
     const conversation: Conversation =
       await this.conversationService.findConversationOfChannel(channelId);
-    // console.log(conversation);
     const conversationUser =
       await this.conversationUserService.findConversationUser(
         conversation.id,
@@ -760,8 +782,17 @@ export class DataService {
   }
 
   async muteUserInChannel(channelId: number, myId: number, body) {
-    console.log('mute');
+    const channel: Channel = await this.channelService.findChannelById(
+      channelId,
+    );
+    if (!channel) return { status: 500, message: 'Channel not found' };
     const chanelUser = await this.findChannelUser(channelId, myId);
+    if (!chanelUser)
+      return { status: 500, message: 'This user is not found in this channel' };
+    if (chanelUser.ban == true) {
+      console.log('This user is already banned');
+      return { status: 500, message: 'This user is already banned' };
+    }
     chanelUser.mute = true;
     chanelUser.timeOfOperation = new Date();
     if (body.time > 0) chanelUser.period = body.time;
@@ -769,11 +800,13 @@ export class DataService {
   }
 
   async banUserInChannel(channelId: number, myId: number, body) {
-    console.log('ban');
     const channel: Channel = await this.channelService.findChannelById(
       channelId,
     );
+    if (!channel) return { status: 500, message: 'Channel not found' };
     const chanelUser = await this.findChannelUser(channelId, myId);
+    if (!chanelUser)
+      return { status: 500, message: 'This user is not found in this channel' };
     chanelUser.ban = true;
     chanelUser.timeOfOperation = new Date();
     if (body.time > 0) chanelUser.period = body.time;
@@ -789,7 +822,10 @@ export class DataService {
     const channel: Channel = await this.channelService.findChannelById(
       channelId,
     );
+    if (!channel) return { status: 500, message: 'Channel not found' };
     const chanelUser = await this.findChannelUser(channelId, myId);
+    if (!chanelUser)
+      return { status: 500, message: 'This user is not found in this channel' };
     await this.messageService.updateHiddenToBeFalse(
       channel.conversation.id,
       myId,
@@ -798,7 +834,13 @@ export class DataService {
   }
 
   async removeMuteUserInChannel(channelId: number, myId: number) {
+    const channel: Channel = await this.channelService.findChannelById(
+      channelId,
+    );
+    if (!channel) return { status: 500, message: 'Channel not found' };
     const chanelUser = await this.findChannelUser(channelId, myId);
+    if (!chanelUser)
+      return { status: 500, message: 'This user is not found in this channel' };
     return await this.channelUserService.removeMute(chanelUser.id);
   }
